Add files via upload
This program is my implementation of the Huffman encoding algorithm to encode a random source S. The aim is to produce an optimal prefix-free encoding f_H, known as the Huffman encoding, which has expected length of f_H bounded by the entropy(S) and entropy(S)+1, i.e. H(S)<=E(f_H)<H(S)+1. Expected length can be thought of as the efficiency of the encoding, with entropy being the randomness/uncertainty of the random source S. Uses trees and lists.