Add files via upload
Algorithm is finished and now appears to run correctly, producing an optimal prefix-free encoding of the given source. Will need to be sharpened in a few areas, particularly when displaying the Huffman encoding. I will also include functions to calculate the expected length of the encoding and the entropy of the source. This in turn enables me to produce further checks on the accuracy of my program, appealing to several theorems than link these properties together. These will be included in the next upload.