Add files via upload
Added functions to calculate the entropy of the random source. This also contains a slight adjustment to attach the Huffman encoding to each node in the printnode method. This allows for easier association with source letter probability and its encoding when calculating the expected length of the Huffman encoding, with the relevant method also included.